Physical/whoop/pyproject.toml
[tool.poetry]
authors = ["Josh Hejka <josh@joshhejka.com>"]
description = "Tools for acquiring and analyzing WHOOP API data."
homepage = "https://github.com/hedgertronic/whoop"
name = "whoop_project"
readme = "README.md"
repository = "https://github.com/hedgertronic/whoop"
version = "0.1.0"

[tool.poetry.dependencies]
Authlib = "^1.0.1"
python = "^3.10"
requests = "^2.28.1"
whoop = "^0.1.0"
pandas = "^2.2.2"
pyairtable = "^2.3.3"
python-dotenv = "0.21.0"

[tool.poetry.group.dev.dependencies]
black = "^22.6.0"
coverage = { version = "^6.4.4", extras = ["toml"] }
flake8 = "^4.0.1"
flake8-bandit = "^3.0.0"
flake8-bugbear = "^22.7.1"
flake8-builtins = "^1.5.3"
flake8-cognitive-complexity = "^0.1.0"
flake8-comprehensions = "^3.10.0"
flake8-docstrings = "^1.6.0"
flake8-eradicate = "^1.3.0"
isort = "^5.10.1"
mypy = "^0.971"
pep8-naming = "^0.13.2"
pre-commit = "^2.20.0"
pylint = "^2.14.5"
pytest = "^7.1.2"
python-dotenv = "^0.21.0"
pyupgrade = "^2.37.3"
tryceratops = "^1.1.0"
types-requests = "^2.28.9"

[tool.coverage.report]
fail_under = 100
show_missing = true

[tool.coverage.run]
omit = ["test*"]

[tool.isort]
lines_after_imports = 2
profile = "black"

[tool.mypy]
pretty = true
show_column_numbers = true
show_error_codes = true
show_error_context = true
warn_unreachable = true

[tool.pylint.messages_control]
ignore-patterns = "test.*?py"

[build-system]
build-backend = "poetry.core.masonry.api"
requires = ["poetry-core>=1.0.0"]
